add sqs sen and sqs consumer(listener) update status

+++ domain/model/src/main/java/co/com/jhompo/model/applicationtype/ApplicationType.java
@@ -1,6 +1,7 @@
 package co.com.jhompo.model.applicationtype;
 import lombok.*;
 
+import java.math.BigDecimal;
 
 
 @Data
@@ -10,8 +11,8 @@ import lombok.*;
 public class ApplicationType {
     private int id;
     private String name;
-    private double minimum_amount;
-    private double maximum_amount;
+    private BigDecimal minimum_amount;
+    private BigDecimal maximum_amount;
     private double interest_rate;
     private boolean automatic_validation;
 }
